chore: sync octobase's update (#32)

+++ y-octo/src/doc/store.rs
@@ -579,6 +579,11 @@ impl DocStore {
                     }
 
                     parent_lock.take();
+                } else {
+                    // if parent not exists, integrate GC node instead
+                    // don't delete it because it may referenced by other nodes
+                    // if all nodes that reference it are deleted, it will merged into one gc node
+                    node = Node::new_gc(node.id(), node.len());
                 }
             }
             Node::GC(item) => {
@@ -599,12 +604,15 @@ impl DocStore {
     }
 
     fn delete_item_inner(delete_set: &mut DeleteSet, item: &Item, parent: Option<&mut YType>) {
+        // 1. mark item as deleted, if item is gced, return
         if !item.delete() {
             return;
         }
 
+        // 2. add it to delete set
         delete_set.add(item.id.client, item.id.clock, item.len());
 
+        // 3. adjust parent length
         if item.parent_sub.is_none() && item.countable() {
             if let Some(parent) = parent {
                 if parent.len != 0 {
@@ -617,11 +625,14 @@ impl DocStore {
 
         match &item.content {
             Content::Type(ty) => {
+                // 4. delete all children
                 if let Some(mut ty) = ty.ty_mut() {
                     // items in ty are all refs, not owned
                     let mut item_ref = ty.start.clone();
                     while let Some(item) = item_ref.get() {
-                        Self::delete_item_inner(delete_set, item, Some(&mut ty));
+                        if !item.deleted() {
+                            Self::delete_item_inner(delete_set, item, Some(&mut ty));
+                        }
 
                         item_ref = item.right.clone();
                     }
@@ -629,7 +640,9 @@ impl DocStore {
                     let map_values = ty.map.values().cloned().collect::<Vec<_>>();
                     for item in map_values {
                         if let Some(item) = item.get() {
-                            Self::delete_item_inner(delete_set, item, Some(&mut ty));
+                            if !item.deleted() {
+                                Self::delete_item_inner(delete_set, item, Some(&mut ty));
+                            }
                         }
                     }
                 }
@@ -715,7 +728,7 @@ impl DocStore {
         diff
     }
 
-    pub fn diff_state_vector(&self, sv: &StateVector) -> JwstCodecResult<Update> {
+    pub fn diff_state_vector(&self, sv: &StateVector, with_pending: bool) -> JwstCodecResult<Update> {
         let update_structs = Self::diff_structs(&self.items, sv)?;
 
         let mut update = Update {
@@ -724,8 +737,10 @@ impl DocStore {
             ..Update::default()
         };
 
-        if let Some(pending) = &self.pending {
-            Update::merge_into(&mut update, [pending.clone()])
+        if with_pending {
+            if let Some(pending) = &self.pending {
+                Update::merge_into(&mut update, [pending.clone()])
+            }
         }
 
         Ok(update)
@@ -808,7 +823,7 @@ impl DocStore {
                         if let Node::Item(item) = items[idx].clone() {
                             let item = unsafe { item.get_unchecked() };
 
-                            if item.id.clock >= end {
+                            if end <= item.id.clock {
                                 break;
                             }
 
@@ -841,7 +856,9 @@ impl DocStore {
         if let Node::Item(item_ref) = items[idx].clone() {
             let item = unsafe { item_ref.get_unchecked() };
 
-            if !item.deleted() {
+            // if replace=true we don't check if the item deleted,
+            // because the parent already delete but children may not delete
+            if !replace && !item.deleted() {
                 return Err(JwstCodecError::Unexpected);
             }
